+++ examples/demo_station/startup_laptop.py
@@ -1,0 +1,9 @@
+import core_tools as ct
+
+# setup logging open database
+ct.configure('./setup_config/ct_config_laptop_full.yaml')
+
+# start in separate processes
+ct.launch_databrowser()
+
+